Refactor server_user_groups to static methods
