+++ server_user_groups.php
@@ -7,10 +7,10 @@
  */
 use PhpMyAdmin\Relation;
 use PhpMyAdmin\Response;
+use PhpMyAdmin\Server\UserGroups;
 use PhpMyAdmin\Server\Users;
 
 require_once 'libraries/common.inc.php';
-require_once 'libraries/server_user_groups.lib.php';
 
 Relation::getRelationsParam();
 if (! $GLOBALS['cfgRelation']['menuswork']) {
@@ -40,37 +40,37 @@ $response->addHTML(Users::getHtmlForSubMenusOnUsersPage('server_user_groups.php'
  * Delete user group
  */
 if (! empty($_REQUEST['deleteUserGroup'])) {
-    PMA_deleteUserGroup($_REQUEST['userGroup']);
+    UserGroups::delete($_REQUEST['userGroup']);
 }
 
 /**
  * Add a new user group
  */
 if (! empty($_REQUEST['addUserGroupSubmit'])) {
-    PMA_editUserGroup($_REQUEST['userGroup'], true);
+    UserGroups::edit($_REQUEST['userGroup'], true);
 }
 
 /**
  * Update a user group
  */
 if (! empty($_REQUEST['editUserGroupSubmit'])) {
-    PMA_editUserGroup($_REQUEST['userGroup']);
+    UserGroups::edit($_REQUEST['userGroup']);
 }
 
 if (isset($_REQUEST['viewUsers'])) {
     // Display users belonging to a user group
-    $response->addHTML(PMA_getHtmlForListingUsersofAGroup($_REQUEST['userGroup']));
+    $response->addHTML(UserGroups::getHtmlForListingUsersofAGroup($_REQUEST['userGroup']));
 }
 
 if (isset($_REQUEST['addUserGroup'])) {
     // Display add user group dialog
-    $response->addHTML(PMA_getHtmlToEditUserGroup());
+    $response->addHTML(UserGroups::getHtmlToEditUserGroup());
 } elseif (isset($_REQUEST['editUserGroup'])) {
     // Display edit user group dialog
-    $response->addHTML(PMA_getHtmlToEditUserGroup($_REQUEST['userGroup']));
+    $response->addHTML(UserGroups::getHtmlToEditUserGroup($_REQUEST['userGroup']));
 } else {
     // Display user groups table
-    $response->addHTML(PMA_getHtmlForUserGroupsTable());
+    $response->addHTML(UserGroups::getHtmlForUserGroupsTable());
 }
 
 $response->addHTML('</div>');
